docs/trading_ai_system_updated/trading_ai_system/ingestion_service/requirements.txt
kafka-python>=2.0.0
ccxt>=4.0.0
tenacity>=8.2.3
python-dotenv>=1.0.0